% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/k-package.R
\docType{package}
\name{k}
\alias{k}
\alias{k-package}
\title{A collection of very shortly named functions for R, in
the spirit of the K programming language.}
\description{
A collection of very shortly named functions for R, in
the spirit of the K programming language.
}

