% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{k}
\alias{k}
\alias{k-package}
\title{A collection of very shortly named functions for R, in
the spirit of the K programming language.}
\description{
A collection of very shortly named functions for R, in
the spirit of the K programming language.
}

